fix: Garmin sync broken — missing unique indexes and bigint overflow
- Create partial unique indexes immediately after dropping old constraint
  (migration was dropping constraint before index creation loop ran)
- Fix garmin_activity_id cast from ::int to ::bigint (values > 2^31)
- Create activity_type_definitions for migrated tags that lack them
- Fix AddData form to show all activity types, not just hardcoded builtins

+++ apps/backend/src/db/activities.ts
@@ -517,7 +517,7 @@ export const deleteGarminActivityWithWrongType = async (
     user,
     `DELETE FROM activities
      WHERE source = 'garmin'
-       AND (data->>'garmin_activity_id')::int = $1
+       AND (data->>'garmin_activity_id')::bigint = $1
        AND activity_type != $2
        AND deleted_at IS NULL
      RETURNING id`,
